WorkforceOS/requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
django-extensions>=3.2,<4.0
django-celery-beat>=2.6,<3.0
drf-spectacular>=0.28,<1.0
psycopg2-binary>=2.9,<3.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-redis>=5.4,<6.0
Pillow>=10.0,<11.0
python-dotenv>=1.0,<2.0
PyJWT>=2.8,<3.0
WeasyPrint>=62.0,<63.0
gunicorn>=21.2,<23.0
boto3>=1.35,<2.0
django-storages>=1.14,<2.0

# Audit history
django-simple-history>=3.5,<4.0

# Monitoring / Observability
sentry-sdk[django]>=1.40,<3.0
django-prometheus>=2.3,<3.0

# Markup
Markdown>=3.6,<4.0

# Export
openpyxl>=3.1,<4.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
pytest-cov>=5.0,<7.0

stripe>=11.0,<15.0

# ConnectOS shared platform layer
connectos-platform @ file:../../ConnectOS/platform
